+++ FactorioCalcBase/dependency_tracker.py
@@ -1,4 +1,3 @@
-from FactorioCalcBase.recipe_class import RecipeClass
 from FactorioCalcBase.oil_processing import OilProcessor
 from FactorioCalcBase.uranium_processing import UraniumProcessor
 from FactorioCalcBase.dependency_dict_common_function import counter_add_dict, merge_item_dicts, \
@@ -12,7 +11,7 @@ class ProcessGeneral:
         if extra_product_rate_dict is None:
             extra_product_rate_dict = {}
         self.extra_product_rate_dict = extra_product_rate_dict
-        self.cannot_process_item_dict = {}
+        self.process_excepts_dict = {}
         self.item_req_dict = {}
         self.recipe_req_dict = {}
         self.construct_dicts()
@@ -22,20 +21,20 @@ class ProcessGeneral:
 
         self.recipe_req_dict = dict_list[0]
         self.item_req_dict = dict_list[1]
-        self.cannot_process_item_dict = dict_list[2]
+        self.process_excepts_dict = dict_list[2]
 
 
 class ProcessExcepts:
-    def __init__(self, cannot_process_dict: dict, extra_product_rate_dict: dict, use_kovarex=False):
+    def __init__(self, process_excepts_dict: dict, extra_product_rate_dict: dict, use_kovarex=False):
         # general -> uranium -> oil
 
-        uranium_obj = UraniumProcessor(cannot_process_dict=cannot_process_dict,
+        uranium_obj = UraniumProcessor(process_excepts_dict=process_excepts_dict,
                                        extra_product_rate_dict=extra_product_rate_dict,
                                        use_kovarex=use_kovarex)
-        uranium_cannot_process_dict = uranium_obj.cannot_process_dict
-        merged_cannot_process_dict = counter_add_dict([cannot_process_dict, uranium_cannot_process_dict])
+        remain_excepts_after_up = uranium_obj.process_excepts_dict
+        merged_excepts_dict = counter_add_dict([process_excepts_dict, remain_excepts_after_up])
 
-        oil_obj = OilProcessor(merged_cannot_process_dict[''], extra_product_rate_dict)
+        oil_obj = OilProcessor(merged_excepts_dict[''], extra_product_rate_dict)
 
         item_dict_list = [uranium_obj.item_needed, oil_obj.item_needed]
         recipe_dict_list = [uranium_obj.recipe_needed, oil_obj.recipe_needed]
@@ -47,7 +46,7 @@ class ProcessExcepts:
 class FullDependency:
     def __init__(self, recipe_name: str, amount: float, extra_product_rate_dict: dict, use_kovarex=False):
         process_general: ProcessGeneral = ProcessGeneral(recipe_name, amount, extra_product_rate_dict)
-        process_excepts: ProcessExcepts = ProcessExcepts(cannot_process_dict=process_general.cannot_process_item_dict,
+        process_excepts: ProcessExcepts = ProcessExcepts(process_excepts_dict=process_general.process_excepts_dict,
                                                          extra_product_rate_dict=extra_product_rate_dict,
                                                          use_kovarex=use_kovarex)
